# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing user accounts and authentication",
        "title": "Account Service API",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/account/create-account": {
            "post": {
                "description": "Creates a new user account with provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Create new account",
                "parameters": [
                    {
                        "description": "Account creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/account/get-account/{user_id}": {
            "get": {
                "description": "Returns account information for specified user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Get user account by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"550e8400-e29b-41d4-a716-446655440000\"",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.GetAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_WebChads_AccountService_internal_models_dtos.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_WebChads_AccountService_internal_models_dtos.CreateAccountRequest": {
            "type": "object",
            "required": [
                "birthdate",
                "firstname",
                "gender",
                "surname"
            ],
            "properties": {
                "birthdate": {
                    "type": "string",
                    "example": "1990-01-01"
                },
                "firstname": {
                    "type": "string",
                    "example": "Иван"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "patronymic": {
                    "type": "string",
                    "example": "Иванович"
                },
                "surname": {
                    "type": "string",
                    "example": "Иванов"
                }
            }
        },
        "github_com_WebChads_AccountService_internal_models_dtos.GetAccountResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 33
                },
                "birthdate": {
                    "type": "string",
                    "example": "1990-01-01T00:00:00Z"
                },
                "firstname": {
                    "type": "string",
                    "example": "Иван"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "patronymic": {
                    "type": "string",
                    "example": "Иванович"
                },
                "surname": {
                    "type": "string",
                    "example": "Иванов"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_WebChads_AccountService_internal_models_dtos.Response": {
            "type": "object",
            "properties": {
                "message": {},
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}